pyproject.toml
[project]
name = "document-simulator"
version = "0.1.0"
description = "Document image augmentation and OCR training system using Augraphy, PaddleOCR, and Stable-Baselines3"
readme = "README.md"
requires-python = ">=3.10,<3.13"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["document", "ocr", "augmentation", "machine-learning", "computer-vision"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Recognition",
]

dependencies = [
    # Document Augmentation
    "augraphy>=8.2.0",

    # OCR
    "paddleocr>=2.7.0",
    "paddlepaddle>=2.6.0",

    # Reinforcement Learning
    "stable-baselines3>=2.3.0",
    "gymnasium>=0.29.0",

    # Computer Vision & Image Processing
    "opencv-python>=4.9.0",
    "pillow>=10.2.0",
    "scikit-image>=0.22.0",
    "albumentations>=1.4.0",

    # Machine Learning & Deep Learning
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "numpy>=1.26.0,<2.0.0",
    "scipy>=1.12.0",

    # Data Handling
    "pandas>=2.2.0",
    "pyarrow>=15.0.0",

    # Utilities
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",
    "rich>=13.7.0",

    # Logging & Monitoring
    "loguru>=0.7.2",
    "tensorboard>=2.16.0",
    "wandb>=0.16.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.2.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.24.0",
]

notebook = [
    "jupyter>=1.0.0",
    "jupyterlab>=4.1.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/document-simulator"
Documentation = "https://github.com/yourusername/document-simulator/docs"
Repository = "https://github.com/yourusername/document-simulator"
Issues = "https://github.com/yourusername/document-simulator/issues"

[project.scripts]
document-simulator = "document_simulator.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.2.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "augraphy.*",
    "paddleocr.*",
    "stable_baselines3.*",
    "cv2.*",
    "albumentations.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--cov=document_simulator",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
    "-v",
]
